Refactor login screen
