+++ lib/android/pages/person_dialog.dart
