+++ lib/android/pages/splash.dart
